[feat] refact && faker
