+++ src/app/new-training/new-training-form/delete-exercise.tsx
